Refactor registry imports to unify command structure and remove deprecated module
