+++ libs/commands/registry/__init__.py
@@ -1,0 +1,6 @@
+"""
+メンバー/チーム操作モジュール
+
+- `libs.commands.registry.member`: メンバー管理
+- `libs.commands.registry.team`: チーム管理
+"""
