+++ libs/commands/registry/team.py
