--- libs/registry/__init__.py
@@ -1,6 +1,0 @@
-"""
-メンバー/チーム操作モジュール
-
-- `libs.registry.member`: メンバー管理
-- `libs.registry.team`: チーム管理
-"""
